pyproject.toml
[project]
name = "stock-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.12",
    "alembic>=1.16.1",
    "alpaca-py>=0.40.1",
    "fastapi>=0.115.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.3.0",
    "openai>=1.86.0",
    "openbb>=4.1.3",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-to-typescript>=2.0.0",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "seaborn>=0.13.2",
    "sqlalchemy>=2.0.41",
    "swarm",
    "uvicorn>=0.32.0",
    "yfinance>=0.2.63",
]

[dependency-groups]
dev = [
    "bandit>=1.8.3",
    "black>=25.1.0",
    "flake8>=7.2.0",
    "isort>=6.0.1",
    "pre-commit>=4.2.0",
    "safety>=3.5.2",
]

[tool.ruff]
line-length = 320
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint]
select = ["E", "F", "I"]  # Enable import sorting (I) along with basic linting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-single-line = false

[tool.uv.sources]
swarm = { git = "https://github.com/openai/swarm.git" }
